refactor(layout): add responsive navigation components
